separate card into separate file and fix the toggle expand bug
